INT-3496: JMX Metrics: Int->Long for all `count`
JIRA: https://jira.spring.io/browse/INT-3496

**Cherry-pick to 4.0.x & 3.0.x**

+++ spring-integration-jmx/src/main/java/org/springframework/integration/monitor/ExponentialMovingAverageRate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2013 the original author or authors.
+ * Copyright 2009-2014 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -92,7 +92,7 @@ public class ExponentialMovingAverageRate {
 	/**
 	 * @return the number of measurements recorded
 	 */
-	public int getCount() {
+	public long getCount() {
 		return rates.getCount();
 	}
 
@@ -107,7 +107,7 @@ public class ExponentialMovingAverageRate {
 	 * @return the mean value
 	 */
 	public double getMean() {
-		int count = rates.getCount();
+		long count = rates.getCount();
 		if (count == 0) {
 			return 0;
 		}
