+++ spring-integration-jmx/src/test/java/org/springframework/integration/monitor/HandlerMonitoringIntegrationTests.java
@@ -1,15 +1,16 @@
 /*
- * Copyright 2009-2010 the original author or authors.
- * 
+ * Copyright 2009-2014 the original author or authors.
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
+
 package org.springframework.integration.monitor;
 
 import static org.junit.Assert.assertEquals;
@@ -36,11 +37,11 @@ public class HandlerMonitoringIntegrationTests {
 	private Service service;
 
 	private IntegrationMBeanExporter messageHandlersMonitor;
-	
+
 	public void setMessageHandlersMonitor(IntegrationMBeanExporter messageHandlersMonitor) {
 		this.messageHandlersMonitor = messageHandlersMonitor;
 	}
-	
+
 	public void setService(Service service) {
 		this.service = service;
 	}
@@ -84,7 +85,7 @@ public class HandlerMonitoringIntegrationTests {
 			channel.send(new GenericMessage<String>("bar"));
 			assertEquals(before + 1, service.getCounter());
 
-			int count = messageHandlersMonitor.getHandlerDuration(monitor).getCount();
+			long count = messageHandlersMonitor.getHandlerDuration(monitor).getCount();
 			assertTrue("No statistics for input channel", count > 0);
 
 		} finally {
@@ -104,7 +105,7 @@ public class HandlerMonitoringIntegrationTests {
 		void execute(String input) throws Exception;
 		int getCounter();
 	}
-	
+
 	public static class SimpleService implements Service {
 		private int counter;
 
@@ -117,7 +118,7 @@ public class HandlerMonitoringIntegrationTests {
 			return counter;
 		}
 	}
-	
+
 	@Aspect
 	public static class HandlerInterceptor {
 		@Before("execution(* *..*Tests*(String)) && args(input)")
